Added fronted for adding elements and navbar

+++ Backend/controllers/product.controller.js
@@ -50,11 +50,14 @@ export const updateProducts = async (req, res) => {
 export const deleteProducts = async (req, res) => {
     // to get the dynamic url part - id
     const { id } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Invalid Product Id" });
+    }
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Product deleted" });
     } catch (error) {
-        res.status(404).json({ success: false, message: "Product not found" });
+        res.status(500).json({ success: false, message: "Server Error" });
     }
 
 };;
